+++ go.mod
@@ -1,0 +1,3 @@
+module gophercises
+
+go 1.15
