+++ quiz/quiz.go
@@ -3,10 +3,8 @@ package main
 import (
 	"bufio"
 	"encoding/csv"
-	"flag"
 	"fmt"
 	"io"
-	"log"
 	"os"
 	"time"
 	"unicode"
@@ -15,27 +13,6 @@ import (
 type problems struct {
 	q string
 	a string
-}
-
-func main() {
-	const defaultName = "problems.csv"
-	const defaultTimeLimit = 30
-	fileName := flag.String("csv", defaultName, "csv file in the format: `question,answer`")
-	timeLimit := flag.Int("limit", defaultTimeLimit, "Integer time limit ")
-	flag.Parse()
-
-	if isValidFileName(*fileName) {
-		arr, err := readFile(*fileName)
-		if err != nil {
-			log.Fatal(err)
-			return
-		}
-		problems := parseQuiz(arr)
-		quiz(problems, *timeLimit)
-	} else {
-		log.Fatal("Invalid file name")
-		return
-	}
 }
 
 func quiz(problems []problems, timeLimit int) {
